[TwigComponent] Fix #3195
This fixes the issue that when relying on the classMap, even components with classes would be loaded as anonymous components, because loading an anonymous template was tried before even checking the classMap.

This fixes the issue by resolving the class map right at the beginning, if possible.

This also allows us to reuse the code that loads the metadata: As we make sure `$name` holds the real config key right at the beginning, we don't have to care anymore about whether the caller provided the internal config key or a class name. The special case for class names can go.

As a side effect, we get a much better error message when attempting to load using an FQCN, because of the code reuse.