[PhpUnit] Drop PHPUnit 9.5 in favor of ^10.5|^11.5
The goal here is to make our CI for Symfony 8.* the greenish as possible. For the moment we have spatie/phpunit-snapshot-assertions and zenstruck/browser that blocks the Symfony 8.0 CI, but spatie/phpunit-snapshot-assertions can be "fixed" when allowing PHPUnit >=10.

The ^10.5|^11.5 constraint is still compatible with PHP 8.1 (our lowest PHP supported version).